Split out FLTK utility functions
These have nothing to do with layout, so let's split them to their own
file.
